Place information elements in a map
Previously, AccessPoint objects contained a data member for every
potential information element (even if the element was not found in the
probe response). These extra data members unnecessarily increased the
size of AccessPoint objects and complicated the D-Bus interface file.
This commit places only the existing information elements in a map,
reducing the size of AccessPoint objects and simplifying the D-Bus
interface file.