Move MainWindow widgets into their own classes
Previously, widgets and their properties were all defined in the
MainWindow's constructor. This commit takes the logic for these widgets
out of the MainWindow class and puts it into dedicated classes for each
widget. For example, all of the widgets that make up the MainWindow's
toolbar are now part of the MainWindowToolbar class.